Extend math onto Quaternion, fix tests that should have broken
The tests include Math in main, which includes it into Object,
making it available to anything that inherits from Object.

This causes singleton methods like Quaternion.rotation to inherit the methods
in the test environment, but not in other environments.

This patch moves the inclusion into the test,
so that it doesn't affect the global namespace,
and then makes the methods available to Quaternion by extending Math onto it.